S12P21A303-87-fix: login filter에 있는 에러 수정

+++ Back/src/main/java/com/ssafy/Dandelion/global/auth/user/CustomUserDetailService.java
@@ -27,7 +27,7 @@ public class CustomUserDetailService implements UserDetailsService {
 
 	private UserDetails createUserDetails(User user) {
 		return CustomUserDetails.builder()
-			.userId(user.getUserId().longValue())
+			.userId(user.getUserId())
 			.email(user.getEmail())
 			.password(user.getPassword())
 			.name(user.getName())
